fix for MSVC

+++ sadl/layer_expand.h
@@ -107,7 +107,7 @@ template<typename T> bool Expand<T>::init(const std::vector<Tensor<T> *> &in)
   if (in[1]->dims().size() != 1)
     return false;
   Dimensions dim;
-  dim.resize(in[1]->size());
+  dim.resize((int)in[1]->size());
   if (!std::is_same<float,T>::value&&in[1]->quantizer!=0) {
       std::cerr << "[ERROR] quantizer on reshape dimensions data layer" << std::endl;
       return false;
